refactor: Restructure project into 'app' and 'tests' directories
This commit restructures the project codebase for better organization and scalability.

Changes include:
- Created a main `app` directory for all application source code.
- Subdirectories within `app` (`core`, `services`, `integrations`) now group modules by functionality.
- All test files moved into a dedicated `tests` directory.
- `__init__.py` files added to all new directories to define them as Python packages.
- All import statements updated المشروعwide to reflect the new module paths.
- Verified that the main simulation (`app/main.py`) and all unit tests pass after restructuring.
- Updated `README.md` to accurately describe the new project structure and provide updated instructions for running code and tests.
