+++ app/core/__init__.py
@@ -1,0 +1,1 @@
+# This file makes 'core' a Python sub-package of 'app'.
