+++ app/integrations/__init__.py
@@ -1,0 +1,1 @@
+# This file makes 'integrations' a Python sub-package of 'app'.
